Split Home into its own js, styled Login screen, added routing
