+++ src/Components/Post/Post.css
@@ -102,7 +102,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  padding-bottom: 5px;
+  padding: 0 0 5px 0;
   top: 10px;
   right: 10px;
 }
